+++ rumqttd/CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - Retained Messages
+- Publish properties in QoS2 publish
 
 ### Security